signed int viddef_t members
The attached patch (against quakeforge git) changes the [con]width,
[con]height, and most importantly the rowbytes members of viddef_t
from unsigned to signed int, like in q2.  This allows for a properly
negative vid.rowbytes which may be needed in, e.g. a DIB sections
windows driver if needed.  Along with it, I changed a few places
where unsigned int is used along with comparisons against the relevant
vid.* members.
One thing I am not 100% sure is the signedness requirements of
d_zrowbytes and d_zwidth: q2 has them as unsigned but I am not sure
whether that is because they are needed as unsigned or it was just an
oversight of the id developers. They do look like they should be OK
as signed int to me, though: comments?

==
Note from Bill Currie: I had to do some extra changes as many
signed/unsigned comparisons were somehow missed.

+++ libs/video/renderer/gl/gl_screen.c
@@ -93,10 +93,8 @@ gl_SCR_ScreenShot (int width, int height)
 	qfglReadPixels (0, 0, vid.width, vid.height, GL_RGB, GL_UNSIGNED_BYTE,
 				    snap);
 
-	w = (vid.width < (unsigned int) width) ? vid.width
-										   : (unsigned int) width;
-	h = (vid.height < (unsigned int) height) ? vid.height
-											 : (unsigned int) height;
+	w = (vid.width < width) ? vid.width : width;
+	h = (vid.height < height) ? vid.height : height;
 
 	fracw = (float) vid.width / (float) w;
 	frach = (float) vid.height / (float) h;
