+++ libs/video/renderer/sw/draw.c
@@ -263,7 +263,7 @@ Draw_Character (int x, int y, unsigned int chr)
 	if (y <= -8)
 		return;							// totally off screen
 
-	if (y > (int) vid.conheight - 8 || x < 0 || x > (int) vid.conwidth - 8)
+	if (y > vid.conheight - 8 || x < 0 || x > vid.conwidth - 8)
 		return;
 	if (chr < 0 || chr > 255)
 		return;
@@ -418,8 +418,8 @@ Draw_Pic (int x, int y, qpic_t *pic)
 	byte       *dest, *source, tbyte;
 	int         v, u;
 
-	if (x < 0 || (unsigned int) (x + pic->width) > vid.conwidth || y < 0 ||
-		(unsigned int) (y + pic->height) > vid.conheight) {
+	if (x < 0 || (x + pic->width) > vid.conwidth
+		|| y < 0 || (y + pic->height) > vid.conheight) {
 		Sys_MaskPrintf (SYS_VID, "Draw_Pic: bad coordinates");
 		Draw_SubPic (x, y, pic, 0, 0, pic->width, pic->height);
 		return;
@@ -477,8 +477,8 @@ Draw_SubPic (int x, int y, qpic_t *pic, int srcx, int srcy, int width,
 	byte       *dest, *source, tbyte;
 	int         u, v;
 
-	if ((x < 0) || (x + width > (int) vid.conwidth)
-		|| (y < 0) || (y + height > (int) vid.conheight)) {
+	if ((x < 0) || (x + width > vid.conwidth)
+		|| (y < 0) || (y + height > vid.conheight)) {
 		Sys_MaskPrintf (SYS_VID, "Draw_SubPic: bad coordinates");
 	}
 	// first, clip to screen
@@ -487,7 +487,7 @@ Draw_SubPic (int x, int y, qpic_t *pic, int srcx, int srcy, int width,
 		width += x;
 		x = 0;
 	}
-	if (x + width > (int) vid.width)
+	if (x + width > vid.width)
 		width = vid.width - x;
 	if (width <= 0)
 		return;
@@ -496,7 +496,7 @@ Draw_SubPic (int x, int y, qpic_t *pic, int srcx, int srcy, int width,
 		height += y;
 		y = 0;
 	}
-	if (y + height > (int) vid.height)
+	if (y + height > vid.height)
 		height = vid.height - y;
 	if (height <= 0)
 		return;
@@ -564,7 +564,7 @@ Draw_ConsoleBackground (int lines, byte alpha)
 		else {
 			f = 0;
 			fstep = 320 * 0x10000 / vid.conwidth;
-			for (x = 0; x < (int) vid.conwidth; x += 4) {
+			for (x = 0; x < vid.conwidth; x += 4) {
 				dest[x] = src[f >> 16];
 				f += fstep;
 				dest[x + 1] = src[f >> 16];
@@ -630,7 +630,7 @@ Draw_TileClear (int x, int y, int w, int h)
 	byte       *psrc;
 	vrect_t     vr;
 
-	CLIP (x, y, w, h, (int) vid.width, (int) vid.height);
+	CLIP (x, y, w, h, vid.width, vid.height);
 
 	r_rectdesc.rect.x = x;
 	r_rectdesc.rect.y = y;
@@ -693,12 +693,12 @@ Draw_Fill (int x, int y, int w, int h, int c)
 	byte       *dest;
 	int         u, v;
 
-	if (x < 0 || x + w > (int) vid.conwidth
-		|| y < 0 || y + h > (int) vid.conheight) {
+	if (x < 0 || x + w > vid.conwidth
+		|| y < 0 || y + h > vid.conheight) {
 		Sys_MaskPrintf (SYS_VID, "Bad Draw_Fill(%d, %d, %d, %d, %c)\n",
 						x, y, w, h, c);
 	}
-	CLIP (x, y, w, h, (int) vid.width, (int) vid.height);
+	CLIP (x, y, w, h, vid.width, vid.height);
 
 	dest = ((byte*)vid.buffer) + y * vid.rowbytes + x;
 	for (v = 0; v < h; v++, dest += vid.rowbytes)
@@ -710,7 +710,7 @@ Draw_Fill (int x, int y, int w, int h, int c)
 void
 Draw_FadeScreen (void)
 {
-	unsigned int x, y;
+	int         x, y;
 	byte       *pbuf;
 
 	VID_UnlockBuffer ();
